Introduce new parallel::interface module
This new module adds an Interface struct that contains all the necessary state to
handle the communication with a ST7920 via parallel bus.

This is done in order to have a better organization of the project and to better
handle the delays in the operations, increasing the number of operations implemented
by the HW agnostic crate and thus reducing the code needed to setup the driver.

This commit also fixes how the bus was handled: by relying solely on the embedded-hal
it's not possible to configure a GPIO to be input or output; when that crate was
added, this fact was not taken into account, and even tho esp_hal::Flex was being used,
it wasn't working because nowhere in the code it was specified if at any given moment
those pins had to be configured as output or as input.

The issue is solved by adding two more traits on top of the embedded-hal ones, that
allow the configuration of the GPIO.