Completely restructure the project
The ST7920 driver now is at the base level crate, the examples are now in a package
and compile to a lib that can be used by the various low-level implementations.

The ESP32 DevKitC examples are now in the esp32 package. There is still one problem,
which is the fact that it can't be compiled with `cargo build -p esp32` as this
ignores the `esp32/rust-toolchain.toml` file which selects the esp channel.

The `st7920::infallible` module got removed as now the error handling is better
organized and the need for that module is gone.

A new trait is added to `st7920::hal::Rng` which ideally should belong to the
examples package, but, due to orphan rules, it would require setting up a feature
for each target on the examples package, which I didn't really like.

Some READMEs are added to describe the new project structure.