Convert the `Clock` trait into an extern function
Both the `Timer` and `Clock` abstraction are ugly to use in my opinion so by
taking inspiration from the `#[global_logger]` machanism (and from C)
I decided to make a single extern function called `_st7920_now` which can
be implemented by the user of the library.

This simplifyies the code by a lot, very cool :)

PS: I've noted that with the tim2 implementation for the STM32, the old
Timer trait could still be used as the main limiting factor when I wrote
the hal module was that while esp32 has `now` integrated in its HAL,
stm32 does not and I didn't know at the time that something like this
could be written.